+++ R/dfm.R
@@ -415,18 +415,20 @@ make_ngram_pattern <- function(features, valuetype, concatenator) {
 
 # create an empty dfm for given features and documents
 make_null_dfm <- function(feature = NULL, document = NULL) {
-    new("dfm",
+    if (is.null(feature)) feature <- character()
+    if (is.null(document)) document <- character()
+    result <- new("dfm",
         as(sparseMatrix(
         i = NULL,
         j = NULL,
-        dims = c(length(document), length(feature)),
-        dimnames = list(docs = document, features = feature)
-    ),
-    "dgCMatrix"))
+        dims = c(length(document), length(feature))
+    ), "dgCMatrix"))
+    set_dfm_dimnames(result) <- list(document, feature)
+    return(result)
 }
 
 # pad dfm with zero-count features
-pad_dfm <- function(x, feature = NULL) {
+pad_dfm <- function(x, feature) {
     feat_pad <- setdiff(feature, featnames(x))
     if (length(feat_pad)) {
         x <- cbind(x, make_null_dfm(feat_pad, docnames(x)))