+++ R/textstat_lexdiv.R
@@ -1,10 +1,14 @@
+# main functions ========
+
 #' Calculate lexical diversity
-#' 
-#' Calculate the lexical diversity or complexity of text(s).
-#' @details \code{textstat_lexdiv} calculates a variety of proposed indices for
-#'   lexical diversity. In the following formulas, \eqn{N} refers to the total
-#'   number of tokens, \eqn{V} to the number of types, and \eqn{f_v(i, N)} to the
-#'   numbers of types occurring \eqn{i} times in a sample of length \eqn{N}.
+#'
+#' Calculate the lexical diversity of text(s).
+#' @details \code{textstat_lexdiv} calculates the lexical diversity of documents
+#'   using a variety of indices.
+#'   
+#' @details In the following formulas, \eqn{N} refers to the total number of
+#'   tokens, \eqn{V} to the number of types, and \eqn{f_v(i, N)} to the numbers
+#'   of types occurring \eqn{i} times in a sample of length \eqn{N}.
 #'   \describe{
 #'      
 #'   \item{\code{"TTR"}:}{The ordinary \emph{Type-Token Ratio}: \deqn{TTR =
@@ -53,12 +57,22 @@
 #'   for the current method (for a dfm) there is no computation on separate
 #'   halves of the text.}
 #'   
+#'   \item{\code{"MATTR"}:}{The Moving-Average Type-Token Ratio (Covington &
+#'   McFall, 2010) calculates TTRs for a moving window of tokens from the first to the last token, computing a TTR for each window. 
+#'   The MATTR is the mean of the TTRs of each window.}
+#'   
+#'   \item{\code{"MSTTR"}:}{Mean Segmental Type-Token Ratio (sometimes referred
+#'   to as \emph{Split TTR}) splits the tokens into segments of the given size,
+#'   TTR for each segment is calculated and the mean of these values returned.
+#'   When this value is < 1.0, it splits the tokens into equal, non-overlapping
+#'   sections of that size.  When this value is > 1, it defines the segments as
+#'   windows of that size. Tokens at the end which do not make a full segment
+#'   are ignored.}
 #'   }
 #'   
-#' @param x an input \link{dfm} or \link{tokens} input object
-#' @param measure a character vector defining the measure to calculate
-#' @param log.base a numeric value defining the base of the logarithm (for
-#'   measures using logs)
+#' @param x an \link{dfm} or \link{tokens} input object for whose documents
+#'   lexical diversity will be computed
+#' @param measure a character vector defining the measure to compute
 #' @param remove_numbers logical; if \code{TRUE} remove features or tokens that
 #'   consist only of numerals (the Unicode "Number" [N] class)
 #' @param remove_punct logical; if \code{TRUE} remove all features or tokens
@@ -69,16 +83,16 @@
 #'   by hyphenation and hyphenation-like characters in between words, e.g.
 #'   "self-storage" becomes two features or tokens "self" and "storage". Default
 #'   is FALSE to preserve such words as is, with the hyphens.
-#' @author Kenneth Benoit and Jiang Wei Lua.  Many of the formulas have been
-#'   re-implemented from functions written by Meik Michalke in the \pkg{koRpus}
+#' @param log.base a numeric value defining the base of the logarithm (for
+#'   measures using logarithms)
+#' @param MATTR_window a numeric value defining the size of the moving window 
+#'   for computation of the Moving-Average Type-Token Ratio (Covington & McFall, 2010)
+#' @param MSTTR_segment a numeric value defining the size of the each segment
+#'   for the computation of the the Mean Segmental Type-Token Ratio (Johnson, 1944)
+#' @param ... for passing arguments to other methods
+#' @author Kenneth Benoit and Jiong Wei Lua.  Many of the formulas have been
+#'   reimplemented from functions written by Meik Michalke in the \pkg{koRpus}
 #'   package.
-#' @note This implements only the static measures of lexical diversity, not more
-#'   complex measures based on windows of text such as the Mean Segmental
-#'   Type-Token Ratio, the Moving-Average Type-Token Ratio (Covington & McFall,
-#'   2010), the MLTD or MLTD-MA (Moving-Average Measure of Textual Lexical
-#'   Diversity) proposed by McCarthy & Jarvis (2010) or Jarvis (no year), or the
-#'   HD-D version of vocd-D (see McCarthy & Jarvis, 2007).  These are available
-#'   from the package \pkg{korRpus}.
 #' @references 
 #'   Covington, Michael A. and Joe D. McFall. 2010. 
 #'   "\href{https://doi.org/10.1080/09296171003643098}{Cutting the Gordian Knot: The 
@@ -110,14 +124,9 @@
 #'   Variable May a Constant Be? Measures of Lexical Richness in Perspective}". \emph{Computers and the 
 #'   Humanities} 32(5): 323--352.
 #'   
-#' @return \code{textstat_lexdiv} returns a data.frame of documents and
-#'   their lexical diversity scores.
+#' @return A data.frame of documents and their lexical diversity scores.
 #' @export
 #' @examples
-#' mydfm <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), verbose = FALSE)
-#' (result <- textstat_lexdiv(mydfm, c("CTTR", "TTR", "U")))
-#' cor(textstat_lexdiv(mydfm, "all")[,-1])
-#' 
 #' txt <- c("Anyway, like I was sayin', shrimp is the fruit of the sea. You can
 #'           barbecue it, boil it, broil it, bake it, saute it.",
 #'          "There's shrimp-kabobs,
@@ -129,108 +138,172 @@
 #'     textstat_lexdiv(measure = c("TTR", "CTTR", "K"))
 #' dfm(txt) %>% 
 #'     textstat_lexdiv(measure = c("TTR", "CTTR", "K"))
-textstat_lexdiv <- function(x, measure = c("all", "TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas"), 
-                            log.base = 10, remove_numbers = TRUE, remove_punct = TRUE,
-                            remove_symbols = TRUE, remove_hyphens = FALSE) {
+#'     
+#' toks <- tokens(corpus_subset(data_corpus_inaugural, Year > 2000))
+#' textstat_lexdiv(toks, c("CTTR", "TTR", "MATTR"), MATTR_window = 100)
+textstat_lexdiv <- function(x,
+                            measure = c("TTR", "C", "R", "CTTR", "U", "S", "K", "D",
+                                        "Vm", "Maas", "MATTR", "MSTTR", "all"),
+                            remove_numbers = TRUE, remove_punct = TRUE,
+                            remove_symbols = TRUE, remove_hyphens = FALSE,
+                            log.base = 10,
+                            MATTR_window = 100L,
+                            MSTTR_segment = 100L,
+                            ...) {
     UseMethod("textstat_lexdiv")
 }
 
 #' @export
-textstat_lexdiv.default <- function(x, measure = c("all", "TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas"), 
-                                log.base = 10, remove_numbers = TRUE, remove_punct = TRUE,
-                                remove_symbols = TRUE, remove_hyphens = FALSE) {
+textstat_lexdiv.default <- function(x, ...) {
     stop(friendly_class_undefined_message(class(x), "textstat_lexdiv"))
 }
 
 #' @export
-textstat_lexdiv.dfm <- 
-    function(x, measure = c("all", "TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas"), 
-             log.base = 10, remove_numbers = TRUE, remove_punct = TRUE,
-             remove_symbols = TRUE, remove_hyphens = FALSE) {
-        
+textstat_lexdiv.dfm <- function(x,
+                                measure = c("TTR", "C", "R", "CTTR", "U", "S", "K", "D",
+                                            "Vm", "Maas", "all"),
+                                remove_numbers = TRUE, remove_punct = TRUE,
+                                remove_symbols = TRUE, remove_hyphens = FALSE,
+                                log.base = 10,
+                                ...) {
+
+    check_dots(list(...))
+    tokens_only_measures <-  c("MATTR", "MSTTR")
+
     x <- as.dfm(x)
     if (!sum(x)) stop(message_error("dfm_empty"))
-    if (remove_hyphens) 
+
+    # special character handling
+    if (remove_hyphens)
         x <- dfm_split_hyphenated_features(x)
-    if (remove_numbers) 
+    if (remove_numbers)
         x <- dfm_remove(x, "^\\p{N}+$", valuetype = "regex")
-    if (remove_punct) 
+    if (remove_punct)
         x <- dfm_remove(x, "^\\p{P}+$", valuetype = "regex")
-    if (remove_symbols) 
+    if (remove_symbols)
         x <- dfm_remove(x, "^\\p{S}+$", valuetype = "regex")
-    
-    if (!sum(x)) 
+
+    if (!sum(x))
         stop(message_error("dfm_empty after removal of numbers, symbols, punctuations, hyphens"))
 
-    measure_option <- c("TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas")
-    if (measure[1] == "all") {
-        measure <- measure_option
-    } else {
-        is_valid <- measure %in% measure_option
-        if (!all(is_valid))
-            stop("Invalid measure(s): ", measure[!is_valid])
-    }
-    
-    compute_lexdiv(x, measure, log.base)
+    # check that no moving average methods have been requested
+    if (any(tokens_only_measures %in% measure))
+        stop("average-based measures are only available for tokens inputs")
+    # match the measures from the function signature
+    available_measures <- as.character(formals()$measure)[-1]
+    # this ensures that a default will choose only the first option
+    measure <- match.arg(measure, choices = available_measures,
+                         several.ok = !missing(measure))
+    # get all measures except "all" if "all" is specified
+    if ("all" %in% measure)
+        measure <- available_measures[!available_measures %in% "all"]
+
+    compute_lexdiv_dfm_stats(x, measure = measure, log.base = log.base)
 }
 
 #' @export
-textstat_lexdiv.tokens <- 
-    function(x, measure = c("all", "TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas"),
-             log.base = 10, remove_numbers = TRUE, remove_punct = TRUE,
-             remove_symbols = TRUE, remove_hyphens = FALSE) {
-    
-    x <- as.tokens(x)
-    if (remove_hyphens)
-        x <- tokens(x, remove_hyphens = TRUE)
-    if (remove_numbers) 
-        x <- tokens(x, remove_numbers = TRUE)
-    if (remove_symbols)
-        x <- tokens(x, remove_symbols = TRUE)
+textstat_lexdiv.tokens <-
+    function(x,
+             measure = c("TTR", "C", "R", "CTTR", "U", "S", "K", "D",
+                         "Vm", "Maas", "MATTR", "MSTTR", "all"),
+             remove_numbers = TRUE, remove_punct = TRUE,
+             remove_symbols = TRUE, remove_hyphens = FALSE,
+             log.base = 10,
+             MATTR_window = 100L,
+             MSTTR_segment = 100L,
+             ...) {
+
+    check_dots(list(...))
+    tokens_only_measures <-  c("MATTR", "MSTTR")
+
+    # additional token handling
+    x <- tokens(x, remove_hyphens = remove_hyphens,
+                remove_numbers = remove_numbers,
+                remove_symbols = remove_symbols)
+
     if (remove_punct) {
-        # this will be replaced with 
+        # this will be replaced with
         # x <- tokens(x, remove_punct = TRUE)
         # when we resolve #1445
         x <- tokens_remove(x, "^\\p{P}+$", valuetype = "regex")
     }
 
-    textstat_lexdiv(dfm(x), measure = measure, log.base = log.base)
+    # get and validate measures
+    available_measures <- as.character(formals()$measure)[-1]
+    # this ensures that a default will choose only the first option
+    measure <- match.arg(measure, choices = available_measures,
+                         several.ok = !missing(measure))
+    # get all measures except "all" if "all" is specified
+    if ("all" %in% measure)
+        measure <- available_measures[!available_measures %in% "all"]
+
+    # which, if any, are tokens-only measures
+    tokens_only_measure_index <- which(measure %in% tokens_only_measures)
+
+    # compute all results - returns NAs for tokens-only measures
+    result <- compute_lexdiv_dfm_stats(dfm(x), measure = measure, log.base = log.base)
+
+    # if any tokens-only measures exist, compute and replace NAs with the reusults
+    # removes the first column which is "documents"
+    if (length(tokens_only_measure_index)) {
+        result[, tokens_only_measure_index + 1] <- # add 1 because of documents column
+            compute_lexdiv_tokens_stats(x,
+                                        measure = measure[tokens_only_measure_index],
+                                        MATTR_window = MATTR_window,
+                                        MSTTR_segment = MSTTR_segment)[, -1]
+    }
+
+    return(result)
+
 }
 
-#' Compute lexical diversity (internal functions)
-#' 
-#' Internal function used in \code{\link{textstat_lexdiv}}.
-#' @name lexdiv
+# internal functions to handle lexdiv statistics for dfm and tokens -------
+
+#' @name compute_lexdiv_stats
+#' @title Compute lexical diversity from a dfm or tokens
+#' @description 
+#' Internal functions used in \code{\link{textstat_lexdiv}}, for computing 
+#' lexical diversity measures on dfms or tokens objects
 #' @param x a \link{dfm} object
 #' @param measure a list of lexical diversity measures.
+#' @return a \code{data.frame} with a \code{document} column containing the
+#'   input document name, followed by columns with the lexical diversity
+#'   statistic, in the order in which they were supplied as the \code{measure}
+#'   argument.
+#' @keywords internal textstat
+NULL
+
+#' @rdname compute_lexdiv_stats
 #' @param log.base a numeric value defining the base of the logarithm (for
 #'   measures using logs)
-#' @return returns a data.frame of documents and their lexical diversity scores.
-compute_lexdiv <- function(x, measure, log.base){
-    
-    n_tokens <- n_types <- TTR <- C <- R <- CTTR <- U <- S <- Maas <- lgV0 <- lgeV0 <- K <- D <- Vm <- NULL
+#' @details \code{compute_lexdiv_dfm_stats} in an internal function that
+#'   computes the lexical diversity measures from a \link{dfm} input.
+compute_lexdiv_dfm_stats <- function(x, measure = NULL, log.base = 10) {
+
+    n_tokens <- n_types <- TTR <- C <- R <- CTTR <- U <- S <- Maas <-
+        lgV0 <- lgeV0 <- K <- D <- Vm <- NULL
     temp <- data.table(n_tokens = ntoken(x), n_types = ntype(x))
-  
+
     if ("TTR" %in% measure)
         temp[, TTR := n_types / n_tokens]
-  
+
     if ("C" %in% measure)
         temp[, C := log(n_types, base = log.base) / log(n_tokens, base = log.base)]
-  
+
     if ("R" %in% measure)
         temp[, R := n_types / sqrt(n_tokens)]
-  
+
     if ("CTTR" %in% measure)
         temp[, CTTR := n_types / sqrt(2 * n_tokens)]
-  
-    if ("U" %in% measure) 
-        temp[, U := log(n_tokens, base = log.base) ^ 2 / 
+
+    if ("U" %in% measure)
+        temp[, U := log(n_tokens, base = log.base) ^ 2 /
                     (log(n_tokens, base = log.base) - log(n_types, base = log.base))]
-  
-    if ("S" %in% measure) 
-        temp[, S := log(log(n_types, base = log.base), base = log.base) / 
+
+    if ("S" %in% measure)
+        temp[, S := log(log(n_types, base = log.base), base = log.base) /
                     log(log(n_tokens, base = log.base), base = log.base)]
-  
+
     # computations for K, D, Vm
     # produces a list of data.frames that will be used for computing the measures
     if (length(intersect(c("K", "D", "Vm"), measure))) {
@@ -243,34 +316,122 @@ compute_lexdiv <- function(x, measure, log.base){
             subset(result, result$i > 0)
       })
     }
-    
+
     if ("K" %in% measure)
-        temp[, K := 10^4 * vapply(ViN, function(y) sum(y$ViN * (y$i / y$n_tokens)^2), numeric(1))]
-    
+        temp[, K := 10 ^ 4 * vapply(ViN, function(y) sum(y$ViN * (y$i / y$n_tokens) ^ 2), numeric(1))]
+
     if ("D" %in% measure)
-        temp[, D := vapply(ViN, 
-                           function(y) sum(y$ViN * (y$i / y$n_tokens) * ((y$i - 1) / (y$n_tokens - 1))), 
+        temp[, D := vapply(ViN,
+                           function(y) sum(y$ViN * (y$i / y$n_tokens) * ( (y$i - 1) / (y$n_tokens - 1)) ),
                            numeric(1))]
-    
-    if ("Vm" %in% measure) 
-        temp[, Vm := vapply(ViN, 
-                            function(y) sqrt( sum(y$ViN * (y$i / y$n_tokens)^2) - 1 / y$n_types[1] ),
+
+    if ("Vm" %in% measure)
+        temp[, Vm := vapply(ViN,
+                            function(y) sqrt( sum(y$ViN * (y$i / y$n_tokens) ^ 2) - 1 / y$n_types[1] ),
                             numeric(1))]
-    
+
     if ("Maas" %in% measure) {
         measure <- c(measure, "lgV0", "lgeV0")
-        temp[, Maas := sqrt((log(n_tokens, base = log.base) - log(n_types, base = log.base)) / 
+        temp[, Maas := sqrt( (log(n_tokens, base = log.base) - log(n_types, base = log.base) ) /
                              log(n_tokens, base = log.base) ^ 2)]
         temp[, lgV0 := log10(n_types) / sqrt(1 - (log10(n_types) / (log10(n_tokens) + 0)) ^ 2)]
         temp[, lgeV0 := log(n_types) / sqrt(1 - (log(n_types) / (log(n_tokens) + 0)) ^ 2)]
     }
-    
+
+    # return missings for tokens-only measures
+    MATTR <- MSTTR <- NULL
+    if ("MATTR" %in% measure) temp[, MATTR := NA]
+    if ("MSTTR" %in% measure) temp[, MSTTR := NA]
+
     result <- data.frame(document = docnames(x), stringsAsFactors = FALSE)
-    result <- cbind(result, as.data.frame(temp[,measure, with = FALSE]))
+    if (length(measure))
+        result <- cbind(result, as.data.frame(temp[, measure, with = FALSE]))
     class(result) <- c("lexdiv", "textstat", "data.frame")
-    rownames(result) <- as.character(seq_len(nrow(result)))
     return(result)
 }
+
+#' @rdname compute_lexdiv_stats
+#' @details \code{compute_lexdiv_tokens_stats} in an internal function that
+#'   computes the lexical diversity measures from a \link{dfm} input.
+#' @param MATTR_window a numeric value defining the size of the moving window 
+#'   for computation of the Moving-Average Type-Token Ratio (Covington & McFall, 2010)
+#' @param MSTTR_segment a numeric value defining the size of the each segment
+#'   for the computation of the the Mean Segmental Type-Token Ratio (Johnson, 1944)
+compute_lexdiv_tokens_stats <- function(x, measure = c("MATTR", "MSTTR"),
+                                     MATTR_window, MSTTR_segment) {
+    measure <- match.arg(measure, several.ok = TRUE)
+    result <- data.frame(document = docnames(x), stringsAsFactors = FALSE)
+
+    if ("MATTR" %in% measure)
+        result[["MATTR"]] <- compute_mattr(x, MATTR_window = MATTR_window)
+
+    if ("MSTTR" %in% measure)
+        result[["MSTTR"]] <- compute_msttr(x, MSTTR_segment = MSTTR_segment)
+
+    # reorder output as originally supplied
+    result <- result[, c("document", measure), drop = FALSE]
+    class(result) <- c("lexdiv", "textstat", "data.frame")
+    return(result)
+}
+
+# specific functions for tokens-only moving average measures -----------
+
+#' Compute the Moving-Average Type-Token Ratio (MATTR)
+#' 
+#' From a tokens object, computes the Moving-Average Type-Token Ratio (MATTR)
+#' from Covington & McFall (2010), averaging all of the sequential moving
+#' windows of tokens of size \code{MATTR_window} across the text, returning the
+#' average as the MATTR.
+#' @param x a \link{tokens} object
+#' @param MATTR_window integer; the size of the moving window for computation of
+#'   TTR, between 1 and the number of tokens of the document
+#' @keywords internal textstat lexdiv
+compute_mattr <- function(x, MATTR_window = 100L) {
+
+    if (MATTR_window < 1)
+        stop("MATTR_window must be positive")
+    if (any(ntoken(x) < MATTR_window)) {
+        MATTR_window <- max(ntoken(x))
+        warning("MATTR_window exceeds some documents' token lengths, resetting to ",
+                MATTR_window)
+    }
+
+    # create each document window as an "ngram"
+    x <- tokens_ngrams(x, n = MATTR_window, concatenator = " ")
+
+    # get a list of TTRs by document
+    temp <- lapply(as.list(x), function(y) textstat_lexdiv(dfm(y), "TTR")[["TTR"]])
+    result <- unlist(lapply(temp, mean))
+    return(result)
+}
+
+#' Compute the Mean Segmental Type-Token Ratio (MSTTR)
+#' 
+#' Compute the Mean Segmental Type-Token Ratio (Johnson 1944) for a tokens input.
+#' @param x input \link{tokens}
+#' @param segment_size the size of the segment
+#' @keywords internal textstat lexdiv
+compute_msttr <- function(x, MSTTR_segment) {
+    if (MSTTR_segment < 1)
+        stop("MSTTR_segment must be positive")
+    if (any(ntoken(x) < MSTTR_segment)) {
+        MSTTR_segment <- max(ntoken(x))
+        warning("MSTTR_segment exceeds some documents' token lengths, resetting to ",
+                MSTTR_segment)
+    }
+
+    x <- tokens_chunk(x, MSTTR_segment)
+    # drop remainder documents shorter than MSTTR_segment
+    x <- x[lengths(x) >= MSTTR_segment]
+
+    temp <- split(textstat_lexdiv(x, measure = "TTR")[["TTR"]],
+                  attr(x, "docvars")[["_docname"]])
+    result <- unlist(lapply(temp, mean))
+    return(result)
+}
+
+
+# additional utility functions ------------
 
 #' Split a dfm's hyphenated features into constituent parts
 #' 
@@ -285,22 +446,24 @@ compute_lexdiv <- function(x, measure, log.base){
 dfm_split_hyphenated_features <- function(x) {
     # the global for matching the hyphens and similar characters
     hyphen_regex <- "^.+\\p{Pd}.+$"
-    
+
     # figure out where the hyphens are
     hyphenated_index <- which(stringi::stri_detect_regex(featnames(x), hyphen_regex))
-    
+
     # return dfm unmodified if no hyphenated features are found
     if (length(hyphenated_index) == 0) return(x)
-    
+
     # split the hyphenated feature names into a list of components
     splitfeatures <- as.list(tokens(featnames(x)[hyphenated_index], remove_hyphens = TRUE))
-    
+
     # efficiently create a new dfm from hyphenated feature name components
     splitdfm <- x[, rep(hyphenated_index, times = lengths(splitfeatures))]
     colnames(splitdfm) <- unlist(splitfeatures, use.names = FALSE)
-    
+
     # combine dfms and suppress duplicated feature name warning
     result <- suppressWarnings(cbind(x[, -hyphenated_index], splitdfm))
+    
     # compress features to combine same-named features
     dfm_compress(result, margin = "features")
-}    
+}
+