+++ tests/testthat/test-textstat.R
@@ -7,7 +7,8 @@ mt <- dfm(toks)
 col <- textstat_collocations(toks, min_count = 1)
 key <- textstat_keyness(mt)
 frq <- textstat_frequency(mt)
-lex <- textstat_lexdiv(mt)
+#Temporary  restrict measures to static measures
+lex <- textstat_lexdiv(mt, measure = c("TTR", "C", "R", "CTTR", "U", "S", "K", "D", "Vm", "Maas"))
 red <- textstat_readability(txt)